autobuild.xml: change 'Unix Makefiles' to Unix Makefiles
So, this allows autobuild to work on my setup. Without this change, I get the error:

    CMake Error: Could not create named generator 'Unix Makefiles'

But feel free to close this if I'm misunderstanding the build setup or something, or
if this breaks things for other people.

I'm running autobuild like this:

    ~/ve/bin/autobuild build -p linux -A 64 --verbose --debug

And I'm using autobuild version 1.1.7, cmake 3.9.3, on Debian Linux.